LPS-170791 Move to osgi-service-tracker-collections package, remove interface and transform methods to static
